+++ .credo.exs
@@ -116,7 +116,7 @@
         ## Refactoring Opportunities
         #
         {Credo.Check.Refactor.CondStatements, []},
-        {Credo.Check.Refactor.CyclomaticComplexity, [max_complexity: 11]},
+        {Credo.Check.Refactor.CyclomaticComplexity, [max_complexity: 12]},
         {Credo.Check.Refactor.FunctionArity, []},
         {Credo.Check.Refactor.LongQuoteBlocks, [max_line_count: 175]},
         # {Credo.Check.Refactor.MapInto, []},